refactoring to make testing easier
