+++ src/clients/transactions/test-client.ts
@@ -1,22 +1,21 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 
+import { CryptoConfig } from 'src/config/crypto-config';
 import winston from 'winston';
 import { Transaction } from '../../types/models';
-import { TransactionReceipt } from '../../types/models/receipt';
 import { ITransactionSubmissionClient } from './base-transaction-client';
+import { GetGasDetails } from './gas';
 
 export class TestTransactionSubmissionClient
   implements ITransactionSubmissionClient
 {
   logger: winston.Logger;
+  cryptoConfig: CryptoConfig;
+  getGasDetails: GetGasDetails;
 
   constructor(logger: winston.Logger) {
     this.logger = logger;
-  }
-
-  quoteTransaction(_transaction: Transaction): Promise<TransactionReceipt> {
-    throw new Error('Method not implemented.');
   }
 
   async sendTransaction(_transaction: Transaction): Promise<any> {
@@ -24,16 +23,9 @@ export class TestTransactionSubmissionClient
     return;
   }
 
-  async getFakePromise(): Promise<any> {
-    return new Promise<number>(resolve => {
-      resolve(-1);
-    });
-  }
-
-  getFromAddress(
+  async getFromAddress(
     _chain: string,
-    _assetSymbol: string
-  ): string | Promise<string> {
+  ): Promise<string> {
     const fakeAddress = '0x123456';
     this.logger.info(`Fake from address: ${fakeAddress}`);
     return fakeAddress;
